pyproject.toml
[project]
name = "simple-mcp-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.0",
    "langchain>=0.3.28",
    "langchain-huggingface>=0.3.1",
    "langchain-mcp-adapters==0.1.0",
    "langgraph>=1.0.1",
    "python-dotenv>=1.2.2",
]
